+++ db.go
@@ -1,0 +1,20 @@
+package gofacto
+
+import (
+	"context"
+	"reflect"
+
+	"github.com/eyo-chen/gofacto/internal/db"
+)
+
+// database is responsible for inserting data into the database
+type database interface {
+	// insert inserts a single data into the database
+	Insert(context.Context, db.InserParams) (interface{}, error)
+
+	// insertList inserts a list of data into the database
+	InsertList(context.Context, db.InserListParams) ([]interface{}, error)
+
+	// GenCustomType generates a non-zero value for custom types
+	GenCustomType(reflect.Type) (interface{}, bool)
+}
